+++ fedot/core/pipelines/pipeline.py
@@ -24,7 +24,7 @@ from fedot.core.repository.tasks import TaskTypesEnum
 from fedot.core.utilities.serializable import Serializable
 from fedot.core.utils import copy_doc
 from fedot.core.visualisation.graph_viz import NodeColorType
-from fedot.core.visualisation.pipeline_specific_utils import PipelineVisualizer
+from fedot.core.visualisation.pipeline_specific_visuals import PipelineVisualizer
 from fedot.preprocessing.preprocessing import DataPreprocessor, update_indices_for_time_series
 
 ERROR_PREFIX = 'Invalid pipeline configuration:'
